BoxText class now works
BoxText class now works mostly as intended, but formatting options applied in the "extra" argument of the initialization section are applied to the whole group. At a minimum, this means that text is drawn with the same fill as the box. If the box fill is white, the text fill is white. If you want black text, then you have to set the fill on the text object after initialization. This is not ideal.