Finalized changes to TSpan coordinates
Fun fact, the dx attribute is the relative offset from the x-coordinate of the previous element, not the parent element. This is so text can be spaced relative to the previous text. Therefore, I had to convert the dx attribute to an x attribute which is no longer relative. I may revert that, but it's working for now, so I'm checking it in.