Add files via upload
I added the boxObj and textObj variables to store the box and text objects created in the BoxText() initialization function. This will help with changes made after initialization to adjust the text.
Also, I added the characterHeightAdjustment variable to the BoxText() class. This is a dirty solution. A more elegant solution would calculate the exact adjustment needed to place the baseline of the bottom row of text such that the text doesn't hang over the bottom of the box.
Finally, I tested the textExtra and boxExtra variables to make sure they worked as intended.