Renamed some variables to be shorter
I renamed the characterWidthMultiplier and characterHeightMultiplier to match typography terms bodyMultiplier and descentMultiplier. The process for grabbing the font-size from the extra variables was improved to grab the font-size from the textExtra variable if it exists and fall back on the font-size defined at the group level if it doesn't exist. I also cleaned up some code to make it easier to read.