Problem with closepath on parent path
There's a problem with the JointedArrow class. The path created has the correct coordinates, but the arrow is filled in like a right triangle. As far as I can tell the subpath is fine; but the parent path is getting filled in even though it is an open path.

It's possible there's nothing wrong with the SVG output. It could be that the browser is closing the path even though I am not telling it to.

Regardless, I think it's probably best to group line commands rather than try to draw the whole arrow with one path command. That way I can reuse the Arrow and Triangle classes.